Publish an actual scala-kernel-api module for Scala 3
scala-kernel-api was published like scala-kernel-api-cross-3.2.1_2.13.10
for Scala 3. This commit also publishes an actual scala-kernel-api_3.2.1
module. That should make the `--shared sh.almond:::scala-kernel-api` option
have an effect when installing a Scala 3 kernel (which should fix some NPEs
seen with Scala 3…) The switch to 'cs bootstrap --hybrid'-generated launchers
in a subsequent commit are non-regression tests for that (most of these tests
don't pass in Scala 3 without the changes here)