Temporarily remove metabrowse stuff
Seems bumping the JDK makes the CI test more things, and unveils
scalameta version mismatches in the class path.

But the recent lack of semanticdb classes from the scalameta Scala.js
artifacts prevents bumping the scalameta version that metabrowse
depends on, for now.

So this temporarily removes the metabrowse stuff from here, so that
we can make the scalameta versions we depend on more consistent, and
unblock releases